adapt to vclib refactoring
