+++ src/main/kotlin/id/walt/auditor/Auditor.kt
@@ -2,7 +2,7 @@ package id.walt.auditor
 
 import id.walt.servicematrix.ServiceProvider
 import id.walt.services.WaltIdService
-import id.walt.vclib.Helpers.toCredential
+import id.walt.vclib.model.toCredential
 import id.walt.vclib.credentials.VerifiablePresentation
 
 import mu.KotlinLogging
