+++ src/main/kotlin/id/walt/auditor/VerificationPolicy.kt
@@ -7,8 +7,8 @@ import id.walt.services.essif.TrustedIssuerClient
 import id.walt.services.key.KeyService
 import id.walt.services.vc.JsonLdCredentialService
 import id.walt.services.vc.JwtCredentialService
-import id.walt.vclib.Helpers.encode
-import id.walt.vclib.VcUtils
+
+
 import id.walt.vclib.credentials.GaiaxCredential
 import id.walt.vclib.credentials.VerifiablePresentation
 import id.walt.vclib.model.VerifiableCredential
@@ -42,7 +42,7 @@ class SignaturePolicy : VerificationPolicy {
         return try {
             log.debug { "is jwt: ${vc.jwt != null}" }
 
-            val issuerDid = VcUtils.getIssuer(vc)!!
+            val issuerDid = vc.issuer!!
 
             try {
                 // Check if key is already in keystore
@@ -91,7 +91,7 @@ class TrustedSchemaRegistryPolicy : VerificationPolicy {
 class TrustedIssuerDidPolicy : VerificationPolicy {
     override val description: String = "Verify by trusted issuer did"
     override fun verify(vc: VerifiableCredential): Boolean {
-        return DidService.loadOrResolveAnyDid(VcUtils.getIssuer(vc)!!) != null
+        return DidService.loadOrResolveAnyDid(vc.issuer!!) != null
     }
 }
 
@@ -104,7 +104,7 @@ class TrustedIssuerRegistryPolicy : VerificationPolicy {
             return true
         }
 
-        val issuerDid = VcUtils.getIssuer(vc)!!
+        val issuerDid = vc.issuer!!
 
         val resolvedIssuerDid =
             DidService.loadOrResolveAnyDid(issuerDid) ?: throw Exception("Could not resolve issuer DID $issuerDid")
@@ -138,7 +138,7 @@ class TrustedIssuerRegistryPolicy : VerificationPolicy {
 class TrustedSubjectDidPolicy : VerificationPolicy {
     override val description: String = "Verify by trusted subject did"
     override fun verify(vc: VerifiableCredential): Boolean {
-        return VcUtils.getSubject(vc)?.let {
+        return vc.subject?.let {
             if (it.isEmpty()) true
             else DidService.loadOrResolveAnyDid(it) != null
         } ?: false
@@ -150,7 +150,7 @@ class IssuanceDateBeforePolicy : VerificationPolicy {
     override fun verify(vc: VerifiableCredential): Boolean {
         return when (vc) {
             is VerifiablePresentation -> true
-            else -> parseDate(VcUtils.getIssuanceDate(vc)).let { it != null && it.before(Date()) }
+            else -> parseDate(vc.issuanceDate).let { it != null && it.before(Date()) }
         }
     }
 }
@@ -160,7 +160,7 @@ class ValidFromBeforePolicy : VerificationPolicy {
     override fun verify(vc: VerifiableCredential): Boolean {
         return when (vc) {
             is VerifiablePresentation -> true
-            else -> parseDate(VcUtils.getValidFrom(vc)).let { it != null && it.before(Date()) }
+            else -> parseDate(vc.validFrom).let { it != null && it.before(Date()) }
         }
     }
 }
@@ -170,7 +170,7 @@ class ExpirationDateAfterPolicy : VerificationPolicy {
     override fun verify(vc: VerifiableCredential): Boolean {
         return when (vc) {
             is VerifiablePresentation -> true
-            else -> parseDate(VcUtils.getExpirationDate(vc)).let { it == null || it.after(Date()) }
+            else -> parseDate(vc.expirationDate).let { it == null || it.after(Date()) }
         }
     }
 }
