+++ src/main/kotlin/id/walt/rest/custodian/CustodianController.kt
@@ -4,7 +4,7 @@ import com.beust.klaxon.Klaxon
 import id.walt.crypto.Key
 import id.walt.crypto.KeyAlgorithm
 import id.walt.custodian.Custodian
-import id.walt.vclib.Helpers.encode
+
 import id.walt.vclib.credentials.VerifiablePresentation
 import id.walt.vclib.model.VerifiableCredential
 import io.javalin.http.Context
