+++ src/main/kotlin/id/walt/services/vcstore/HKVVcStoreService.kt
@@ -2,8 +2,8 @@ package id.walt.services.vcstore
 
 import id.walt.services.context.ContextManager
 import id.walt.services.hkvstore.HKVKey
-import id.walt.vclib.Helpers.encode
-import id.walt.vclib.Helpers.toCredential
+
+import id.walt.vclib.model.toCredential
 import id.walt.vclib.model.VerifiableCredential
