+++ src/test/kotlin/id/walt/services/jwt/JwtServiceTest.kt
@@ -17,9 +17,9 @@ import id.walt.services.did.DidService
 import id.walt.services.key.KeyService
 import id.walt.services.keystore.KeyType
 import id.walt.test.RESOURCES_PATH
-import id.walt.vclib.Helpers.encode
-import id.walt.vclib.Helpers.toCredential
-import id.walt.vclib.Helpers.toMap
+
+import id.walt.vclib.model.toCredential
+
 import id.walt.vclib.credentials.Europass
 import io.kotest.core.spec.style.AnnotationSpec
 import io.kotest.matchers.shouldBe
